+++ demos/54-aftereffects-test/index.js
@@ -21,7 +21,7 @@
   let GUP = gup();
 
   window.addEventListener("load", function () {
-    var isDarkmode = detectDarkMode(GUP);
+    /* var isDarkmode = detectDarkMode(GUP);
     var params = new Params(GUP);
 
     // All config params are optional.
@@ -424,9 +424,20 @@
         pb.drawConfig.circle.color = "rgba(0,0,0,0)";
       }
     };
+  */
 
     // ------ NEW
+    var effectsConfig = {
+      color: "#ff0000",
+      colorAlpha: 0.95
+    };
+
+    var getOpacityPct = function () {
+      return "" + effectsConfig.colorAlpha * 100.0 + "%";
+    };
+
     var getBackdropFilterString = function () {
+      console.log("getOpacityPct", getOpacityPct());
       var buffer = [];
       // buffer.push("blur(5px)");
       buffer.push("invert(80%)");
@@ -439,11 +450,13 @@
       // buffer.push("hue-rotate(120deg)");
       // buffer.push("invert(70%)");
       // buffer.push("opacity(20%)");
+      buffer.push("opacity(" + getOpacityPct() + ")");
       // buffer.push("sepia(90%)");
       buffer.push("saturate(100%)");
 
       return buffer.join(" ");
     };
+
     var canvas = document.getElementById("my-canvas");
     var canvasParent = canvas.parentElement;
     var effectsNode = document.createElement("div");
@@ -453,63 +466,78 @@
     effectsNode.style["width"] = "100%";
     effectsNode.style["height"] = "100%";
     // effectsNode.style["background-color"] = "rgba(255,0,0,0.5)";
+    // effectsNode.style["background-color"] = effectsConfig.color;
+    // effectsNode.style["opacity"] = effectsConfig.colorAlpha;
     effectsNode.style["pointer-events"] = "none";
-    effectsNode.style["backdrop-filter"] = getBackdropFilterString();
+    // effectsNode.style["backdrop-filter"] = getBackdropFilterString();
     // effectsNode.style["backdrop-filter"] = "blur(5px)";
     // effectsNode.style["backdrop-filter"] = "invert(80%)";
     // effectsNode.style["backdrop-filter"] = "sepia(90%)";
 
     canvasParent.appendChild(effectsNode);
+
+    var updateBackdropFilter = function () {
+      effectsNode.style["backdrop-filter"] = getBackdropFilterString();
+    };
+    updateBackdropFilter();
     // ----- /NEW
 
-    // +---------------------------------------------------------------------------------
-    // | Initialize dat.gui
-    // +-------------------------------
-    {
-      var gui = pb.createGUI();
-      // prettier-ignore
-      gui.add(config, "numCircles").min(1).max(10).step(1).onChange( function() { reinit(); rebuildMetaballs(); pb.redraw(); } ).name('numCircles').title("Number of circles.");
-      // prettier-ignore
-      gui.add(config, "metaRadiusAddon").min(0).max(100).step(1).onChange(function () {
-        // rebuildContainingCircles();
-        rebuildMetaballs();
-        pb.redraw();
-      }).name("metaRadiusAddon").title("The metaball connection factor.");
-      // prettier-ignore
-      gui.add(config, "drawCircles").onChange( function() { toggleCircleVisibility(); pb.redraw(); } ).name('drawCircles').title("Draw circles?");
-      // prettier-ignore
-      gui.add(config, "drawContainingCircles").onChange( function() { pb.redraw(); } ).name('drawContainingCircles').title("Draw containing circles?");
-      // prettier-ignore
-      gui.add(config, "drawInverseCircles").onChange( function() { pb.redraw(); } ).name('drawInverseCircles').title("Draw inverse circles at intersection points?");
-      // prettier-ignore
-      gui.add(config, "drawCircleNumbers").onChange( function() { pb.redraw(); } ).name('drawCircleNumbers').title("Draw circle numbers?");
-      // prettier-ignore
-      gui.add(config, "drawOuterHull").onChange( function() { pb.redraw(); } ).name('drawOuterHull').title("Draw outer hull?");
-      // prettier-ignore
-      gui.add(config, "epsilonPathDetect").min(0.0).max(1.0).onChange( function() { rebuildMetaballs(); pb.redraw(); } ).name('epsilonPathDetect').title("Which epslion to use for connected path detection.");
-      // prettier-ignore
-      gui.add(config, "readme").name('readme').title("Display this demo's readme.");
+    globalThis.demoInitializationObserver
+      .waitForInitialized()
+      .then(function (initializedPB) {
+        console.log("initializedPB", initializedPB);
+        var pb = initializedPB;
+        var gui = pb.getGUI();
 
-      var tmpConf = {
-        color: "#ff0000",
-        colorAlpha: 0.0
-      };
-      gui.addColor(tmpConf, "color").onChange(function (newValue) {
-        console.log("New value (0)", newValue);
+        // +---------------------------------------------------------------------------------
+        // | Initialize dat.gui
+        // +-------------------------------
+        {
+          //     var gui = pb.createGUI();
+          //     // prettier-ignore
+          //     gui.add(config, "numCircles").min(1).max(10).step(1).onChange( function() { reinit(); rebuildMetaballs(); pb.redraw(); } ).name('numCircles').title("Number of circles.");
+          //     // prettier-ignore
+          //     gui.add(config, "metaRadiusAddon").min(0).max(100).step(1).onChange(function () {
+          //   // rebuildContainingCircles();
+          //   rebuildMetaballs();
+          //   pb.redraw();
+          // }).name("metaRadiusAddon").title("The metaball connection factor.");
+          //     // prettier-ignore
+          //     gui.add(config, "drawCircles").onChange( function() { toggleCircleVisibility(); pb.redraw(); } ).name('drawCircles').title("Draw circles?");
+          //     // prettier-ignore
+          //     gui.add(config, "drawContainingCircles").onChange( function() { pb.redraw(); } ).name('drawContainingCircles').title("Draw containing circles?");
+          //     // prettier-ignore
+          //     gui.add(config, "drawInverseCircles").onChange( function() { pb.redraw(); } ).name('drawInverseCircles').title("Draw inverse circles at intersection points?");
+          //     // prettier-ignore
+          //     gui.add(config, "drawCircleNumbers").onChange( function() { pb.redraw(); } ).name('drawCircleNumbers').title("Draw circle numbers?");
+          //     // prettier-ignore
+          //     gui.add(config, "drawOuterHull").onChange( function() { pb.redraw(); } ).name('drawOuterHull').title("Draw outer hull?");
+          //     // prettier-ignore
+          //     gui.add(config, "epsilonPathDetect").min(0.0).max(1.0).onChange( function() { rebuildMetaballs(); pb.redraw(); } ).name('epsilonPathDetect').title("Which epslion to use for connected path detection.");
+          //     // prettier-ignore
+          //     gui.add(config, "readme").name('readme').title("Display this demo's readme.");
+
+          gui.addColor(effectsConfig, "color").onChange(function (newValue) {
+            console.log("New value (0)", newValue);
+          });
+          gui
+            .addColorWithAlpha(effectsConfig, "color", "colorAlpha")
+            .onChange(function (newColorValue, newAlphaValue) {
+              console.log("New value (1)", newColorValue, newAlphaValue);
+              updateBackdropFilter();
+            })
+            .name("TEST")
+            .title("test");
+          // globalThis.myGui = gui;
+        }
+
+        // pb.config.postDraw = redraw;
+        // init();
+        // rebuildMetaballs();
+        // pb.redraw();
+      })
+      .catch(function (error) {
+        console.error("Failed to retrieve PB instance.", error);
       });
-      gui
-        .addColorWithAlpha(tmpConf, "color", "colorAlpha")
-        .onChange(function (newColorValue, newAlphaValue) {
-          console.log("New value (1)", newColorValue, newAlphaValue);
-        })
-        .name("TEST")
-        .title("test");
-      globalThis.myGui = gui;
-    }
-
-    pb.config.postDraw = redraw;
-    init();
-    rebuildMetaballs();
-    pb.redraw();
   });
 })(window);